change to int

+++ src/run.py
@@ -228,7 +228,7 @@ def train_step(images, disc_updates):
     """
 
     for _ in range(disc_updates):
-        noise = tf.random.normal([BATCH_SIZE/2, NOISE_DIM])
+        noise = tf.random.normal([int(BATCH_SIZE/2), NOISE_DIM])
         with tf.GradientTape() as disc_tape:
             generated_images = generator(noise, training=True)
 
@@ -245,7 +245,7 @@ def train_step(images, disc_updates):
         discriminator_optimizer.apply_gradients(
             zip(gradients_of_discriminator, discriminator.trainable_variables))
 
-    noise = tf.random.normal([BATCH_SIZE/2, NOISE_DIM])
+    noise = tf.random.normal([int(BATCH_SIZE/2), NOISE_DIM])
     with tf.GradientTape() as gen_tape:
         generated_images = generator(noise, training=True)
